requirements.txt
# ElevenLabs Connect - Python Dependencies
#
# Install with: pip install -r requirements.txt

# HTTP client for API calls
requests>=2.28.0

# Environment variable management
python-dotenv>=1.0.0

# WebSocket client for real-time conversations
websockets>=12.0

# Optional: For running tests
pytest>=7.0.0
pytest-asyncio>=0.21.0
